# frontend/openapi.yaml
# 백엔드 API 계약서 (OpenAPI 3.0).
#
# 백엔드가 스펙을 노출하면(예: /v3/api-docs.yaml) 그 내용으로 이 파일을 갱신하고
# `pnpm gen:api` 로 src/api/generated 를 재생성한다.
#
# 주의: axios baseURL(src/api/mutator/custom-instance.ts)이 이미 `/api` 를
# 포함하므로, 아래 경로들은 실제 백엔드 컨트롤러의 `/api` 접두사를 뗀 상태로
# 적는다 (예: 백엔드 `/api/products` -> 여기서는 `/products`).
# `/health` 만 예외 — 백엔드에 `/api` 접두사가 없는 별도 엔드포인트라 그대로 둔다
# (기존부터 있던 것이라 이번 작업 범위 밖).
openapi: 3.0.3
info:
  title: KIRIN API
  version: 0.1.0
paths:
  /health:
    get:
      operationId: getHealth
      tags:
        - health
      summary: 헬스 체크
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /auth/login:
    post:
      operationId: login
      tags:
        - auth
      summary: 로그인
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: 로그인 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberResponse"
        default:
          $ref: "#/components/responses/Error"

  /auth/guest-login:
    post:
      operationId: guestLogin
      tags:
        - auth
      summary: 게스트 로그인 (인증정보 없이 호출)
      responses:
        "201":
          description: 게스트 세션 발급
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberResponse"
        default:
          $ref: "#/components/responses/Error"

  /auth/logout:
    post:
      operationId: logout
      tags:
        - auth
      summary: 로그아웃
      responses:
        "204":
          description: 세션 무효화됨
        default:
          $ref: "#/components/responses/Error"

  /members:
    post:
      operationId: signUp
      tags:
        - members
      summary: 회원가입
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignUpRequest"
      responses:
        "201":
          description: 가입 완료
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberResponse"
        default:
          $ref: "#/components/responses/Error"

  /members/me:
    get:
      operationId: getMyInfo
      tags:
        - members
      summary: 내 정보 조회
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberResponse"
        default:
          $ref: "#/components/responses/Error"

  /products:
    post:
      operationId: registerProduct
      tags:
        - products
      summary: 물품 등록
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterProductRequest"
      responses:
        "200":
          description: 등록된 물품
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        default:
          $ref: "#/components/responses/Error"
    get:
      operationId: getProducts
      tags:
        - products
      summary: 상태별 물품 목록 조회 (기본 SELLING)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ProductStatus"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
        default:
          $ref: "#/components/responses/Error"

  /products/me:
    get:
      operationId: getMyProducts
      tags:
        - products
      summary: 내 물품 목록 조회
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ProductStatus"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
        default:
          $ref: "#/components/responses/Error"

  /products/{productId}:
    get:
      operationId: getProduct
      tags:
        - products
      summary: 물품 상세 조회
      parameters:
        - $ref: "#/components/parameters/ProductId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        default:
          $ref: "#/components/responses/Error"

  /products/{productId}/locker-reservation:
    post:
      operationId: reserveLocker
      tags:
        - products
      summary: 사물함 예약
      parameters:
        - $ref: "#/components/parameters/ProductId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReserveLockerRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        default:
          $ref: "#/components/responses/Error"

  /products/{productId}/locker-reservation/cancel:
    post:
      operationId: cancelLockerReservation
      tags:
        - products
      summary: 사물함 예약 취소
      parameters:
        - $ref: "#/components/parameters/ProductId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        default:
          $ref: "#/components/responses/Error"

  /products/{productId}/deposit-start:
    post:
      operationId: startDeposit
      tags:
        - products
      summary: 입고 시작
      parameters:
        - $ref: "#/components/parameters/ProductId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        default:
          $ref: "#/components/responses/Error"

  /products/{productId}/deposit-complete:
    post:
      operationId: completeDeposit
      tags:
        - products
      summary: 입고 완료 (판매 시작)
      parameters:
        - $ref: "#/components/parameters/ProductId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        default:
          $ref: "#/components/responses/Error"

  /products/{productId}/recovery-start:
    post:
      operationId: startRecovery
      tags:
        - products
      summary: 회수 시작
      parameters:
        - $ref: "#/components/parameters/ProductId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        default:
          $ref: "#/components/responses/Error"

  /products/{productId}/recovery-complete:
    post:
      operationId: completeRecovery
      tags:
        - products
      summary: 회수 완료
      parameters:
        - $ref: "#/components/parameters/ProductId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        default:
          $ref: "#/components/responses/Error"

  /lockers:
    get:
      operationId: getLockers
      tags:
        - lockers
      summary: 사물함 전체 목록 조회
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LockerListResponse"
        default:
          $ref: "#/components/responses/Error"

  /lockers/{lockerId}/lock-status:
    get:
      operationId: getLockStatus
      tags:
        - lockers
      summary: 사물함 잠금 상태 조회
      parameters:
        - $ref: "#/components/parameters/LockerId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LockerLockStatusResponse"
        default:
          $ref: "#/components/responses/Error"
    patch:
      operationId: changeLockStatus
      tags:
        - lockers
      summary: 사물함 잠금 상태 변경 (판매자 본인만)
      parameters:
        - $ref: "#/components/parameters/LockerId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeLockStatusRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LockerLockStatusResponse"
        default:
          $ref: "#/components/responses/Error"

  /lockers/{lockerId}/product:
    get:
      operationId: getProductByLocker
      tags:
        - lockers
      summary: 사물함 QR 스캔으로 그 안에 있는 물품 조회 (로그인 불필요)
      parameters:
        - $ref: "#/components/parameters/LockerId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        default:
          $ref: "#/components/responses/Error"

  /lockers/{lockerId}/reset:
    post:
      operationId: resetLocker
      tags:
        - lockers
      summary: 사물함 초기화 (관리자용, 잠김+비어있음으로 강제 복구)
      parameters:
        - $ref: "#/components/parameters/LockerId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LockerSummaryResponse"
        default:
          $ref: "#/components/responses/Error"

  /transactions:
    post:
      operationId: purchaseProduct
      tags:
        - transactions
      summary: 물품 구매 (결제 승인 포함, 거래 생성)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseProductRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
        default:
          $ref: "#/components/responses/Error"

  /transactions/{transactionId}:
    get:
      operationId: getTransaction
      tags:
        - transactions
      summary: 거래 상세 조회
      parameters:
        - $ref: "#/components/parameters/TransactionId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
        default:
          $ref: "#/components/responses/Error"

  /transactions/{transactionId}/pickup-complete:
    post:
      operationId: completePickup
      tags:
        - transactions
      summary: 수령 완료 처리
      parameters:
        - $ref: "#/components/parameters/TransactionId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
        default:
          $ref: "#/components/responses/Error"

  /payments/confirm:
    post:
      operationId: confirmPayment
      tags:
        - payments
      summary: 토스 결제 승인 (단독 호출용 — 구매 플로우는 /transactions 로 충분)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentConfirmRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentConfirmResponse"
        default:
          $ref: "#/components/responses/Error"

  /images:
    post:
      operationId: uploadImage
      tags:
        - images
      summary: 이미지 업로드 (서버 로컬 디스크 저장)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: 업로드 완료
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageUploadResponse"
        default:
          $ref: "#/components/responses/Error"

components:
  parameters:
    ProductId:
      name: productId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    LockerId:
      name: lockerId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    TransactionId:
      name: transactionId
      in: path
      required: true
      schema:
        type: integer
        format: int64

  responses:
    Error:
      description: 에러 응답
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  schemas:
    ErrorResponse:
      type: object
      required: [status, code, message, path, timestamp]
      properties:
        status:
          type: integer
        code:
          type: string
        message:
          type: string
        path:
          type: string
        timestamp:
          type: string
          format: date-time

    MemberType:
      type: string
      enum: [REGISTERED, GUEST]

    MemberResponse:
      type: object
      required: [id, loginId, nickname, memberType]
      properties:
        id:
          type: integer
          format: int64
        loginId:
          type: string
        nickname:
          type: string
        memberType:
          $ref: "#/components/schemas/MemberType"

    LoginRequest:
      type: object
      required: [loginId, password]
      properties:
        loginId:
          type: string
        password:
          type: string

    SignUpRequest:
      type: object
      required: [loginId, password, nickname]
      properties:
        loginId:
          type: string
          minLength: 4
          maxLength: 20
        password:
          type: string
          minLength: 8
          maxLength: 64
        nickname:
          type: string
          minLength: 2
          maxLength: 30

    ProductStatus:
      type: string
      enum: [PREPARING, RESERVED, SELLING, SOLD, EXPIRED]

    RegisterProductRequest:
      type: object
      required: [name, price]
      properties:
        name:
          type: string
          maxLength: 100
        price:
          type: integer
          format: int64
          minimum: 1000
          maximum: 1000000000
        description:
          type: string
          maxLength: 1000
          nullable: true
        imageUrls:
          type: array
          maxItems: 10
          nullable: true
          items:
            type: string
            maxLength: 500

    ReserveLockerRequest:
      type: object
      required: [lockerId]
      properties:
        lockerId:
          type: integer
          format: int64

    ProductResponse:
      type: object
      required:
        [
          productId,
          lockerId,
          name,
          price,
          description,
          imageUrls,
          sellerName,
          status,
          createdAt,
          reservedAt,
          reservationExpiresAt,
          depositStartedAt,
          sellingStartedAt,
          sellingExpiresAt,
          recoveryStartedAt,
        ]
      properties:
        productId:
          type: integer
          format: int64
        lockerId:
          type: integer
          format: int64
          nullable: true
        name:
          type: string
        price:
          type: integer
          format: int64
        description:
          type: string
          nullable: true
        imageUrls:
          type: array
          items:
            type: string
        sellerName:
          type: string
        status:
          $ref: "#/components/schemas/ProductStatus"
        createdAt:
          type: string
          format: date-time
        reservedAt:
          type: string
          format: date-time
          nullable: true
        reservationExpiresAt:
          type: string
          format: date-time
          nullable: true
        depositStartedAt:
          type: string
          format: date-time
          nullable: true
        sellingStartedAt:
          type: string
          format: date-time
          nullable: true
        sellingExpiresAt:
          type: string
          format: date-time
          nullable: true
        recoveryStartedAt:
          type: string
          format: date-time
          nullable: true
        sellerCompletedSalesCount:
          type: integer
          format: int64
          nullable: true
          description: 판매자가 지금까지 완료한 판매 건수. 상품 상세 화면에서만 채워진다.

    ProductSummaryResponse:
      type: object
      required: [productId, lockerId, name, price, imageUrl, status]
      properties:
        productId:
          type: integer
          format: int64
        lockerId:
          type: integer
          format: int64
          nullable: true
        name:
          type: string
        price:
          type: integer
          format: int64
        imageUrl:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/ProductStatus"

    ProductListResponse:
      type: object
      required: [products]
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/ProductSummaryResponse"

    LockStatus:
      type: string
      enum: [LOCKED, UNLOCKED]

    UsageStatus:
      type: string
      enum: [AVAILABLE, RESERVED, OCCUPIED]

    LockerSummaryResponse:
      type: object
      required:
        [
          lockerId,
          lockStatus,
          usageStatus,
          isMine,
          reservationExpiresAt,
          sellingExpiresAt,
        ]
      properties:
        lockerId:
          type: integer
          format: int64
        lockStatus:
          $ref: "#/components/schemas/LockStatus"
        usageStatus:
          $ref: "#/components/schemas/UsageStatus"
        isMine:
          type: boolean
          description: >-
            로그인한 회원이 이 사물함을 점유 중인 물품의 판매자인지 여부.
            false면(비로그인 포함) usageStatus는 RESERVED/OCCUPIED를
            구분하지 않고 전부 OCCUPIED로 표시된다.
        reservationExpiresAt:
          type: string
          format: date-time
          nullable: true
          description: isMine && usageStatus=RESERVED일 때만 값이 있다.
        sellingExpiresAt:
          type: string
          format: date-time
          nullable: true
          description: isMine && 판매중(usageStatus=OCCUPIED)일 때만 값이 있다.

    LockerListResponse:
      type: object
      required: [lockers]
      properties:
        lockers:
          type: array
          items:
            $ref: "#/components/schemas/LockerSummaryResponse"

    LockerLockStatusResponse:
      type: object
      required: [lockerId, lockStatus]
      properties:
        lockerId:
          type: integer
          format: int64
        lockStatus:
          $ref: "#/components/schemas/LockStatus"

    ChangeLockStatusRequest:
      type: object
      required: [lockStatus]
      properties:
        lockStatus:
          $ref: "#/components/schemas/LockStatus"

    TransactionStatus:
      type: string
      enum: [PAID, DONE]

    PurchaseProductRequest:
      type: object
      required: [productId, paymentKey, orderId, amount]
      properties:
        productId:
          type: integer
          format: int64
        paymentKey:
          type: string
        orderId:
          type: string
        amount:
          type: integer
          format: int64

    TransactionResponse:
      type: object
      required:
        [
          transactionId,
          productId,
          lockerId,
          buyerName,
          price,
          status,
          orderId,
          approvedAt,
          createdAt,
        ]
      properties:
        transactionId:
          type: integer
          format: int64
        productId:
          type: integer
          format: int64
        lockerId:
          type: integer
          format: int64
          nullable: true
        buyerName:
          type: string
        price:
          type: integer
          format: int64
        status:
          $ref: "#/components/schemas/TransactionStatus"
        orderId:
          type: string
        approvedAt:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time

    PaymentConfirmRequest:
      type: object
      required: [paymentKey, orderId, amount]
      properties:
        paymentKey:
          type: string
        orderId:
          type: string
        amount:
          type: integer
          format: int64

    PaymentConfirmResponse:
      type: object
      required: [paymentKey, orderId, status, amount, approvedAt]
      properties:
        paymentKey:
          type: string
        orderId:
          type: string
        status:
          type: string
        amount:
          type: integer
          format: int64
        approvedAt:
          type: string
          nullable: true

    ImageUploadResponse:
      type: object
      required: [imageUrl]
      properties:
        imageUrl:
          type: string